requirements.txt
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
streamlit>=1.34.0

